+++ regimes/it/scenarios.go
@@ -3,6 +3,7 @@ package it
 import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -16,8 +17,8 @@ var invoiceScenarios = &tax.ScenarioSet{
 		// **** MESSAGES ****
 		{
 			Tags: []cbc.Key{tax.TagReverseCharge},
-			Note: &cbc.Note{
-				Key:  cbc.NoteKeyLegal,
+			Note: &tax.ScenarioNote{
+				Key:  org.NoteKeyLegal,
 				Src:  tax.TagReverseCharge,
 				Text: "Reverse Charge / Inversione del soggetto passivo",
 			},