+++ CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - `gr-mydata`: added `gr-mydata-other-tax` extension to set the category of other taxes in charges.
+- `untdid`: 4451 - `untdid-text-subject` (Text Subject Qualifer) catalogue.
+
+### Changed
+
+- `org`: Moved `cbc.Note` to `org.Note` in order to include extensions.
+- `tax`: using `ScenarioNote` instead of `cbc.Note` to avoid cyclic dependency issues.
 
 ## [v0.207.0] - 2024-12-12